Move strip_trailing_slash to util module
