+++ crates/iceberg/src/spec/table_properties.rs
@@ -21,6 +21,7 @@ use std::str::FromStr;
 
 use crate::compression::CompressionCodec;
 use crate::error::{Error, ErrorKind, Result};
+use crate::util::location;
 
 fn parse_property<T: FromStr>(
     properties: &HashMap<String, String>,
@@ -40,18 +41,6 @@ where
     })
 }
 
-/// Strips trailing slashes from a location, preserving a bare URI scheme root
-pub(crate) fn strip_trailing_slash(path: &str) -> &str {
-    let mut path = path;
-    while !path.ends_with("://") {
-        let Some(stripped) = path.strip_suffix('/') else {
-            break;
-        };
-        path = stripped;
-    }
-    path
-}
-
 fn parse_location_property(
     properties: &HashMap<String, String>,
     key: &str,
@@ -66,7 +55,7 @@ fn parse_location_property(
                 ));
             }
 
-            Ok(strip_trailing_slash(path).to_string())
+            Ok(location::strip_trailing_slash(path).to_string())
         })
         .transpose()
 }
@@ -460,6 +449,7 @@ impl TryFrom<&HashMap<String, String>> for TableProperties {
 mod tests {
     use super::*;
     use crate::compression::CompressionCodec;
+    use crate::util::location::strip_trailing_slash;
 
     #[test]
     fn test_table_properties_default() {
